5.2/BUGFIX/floating point accuracy improved

+++ 5-2_getfloat/ft_getfloat.c
@@ -32,7 +32,7 @@ int	ft_getfloat(float *pn, FILE *fd)
 			*pn = 10 * *pn + (c - '0');
 
 		if (c == '.') {
-			float	div;
+			double	div;
 			div = 1;
 			while (isdigit((c = getc(fd)))) {
 				div *= 0.1;
